Created a whole new level system with 8 different levels.
The level system is composed of a new class called Level and an Enum
called LevelNum. Changes to NachoPanel include commenting out testRun()
in actionPerformed method and adding a new call to runGame() in its
place. RunGame runs the current level. A new instance variable for Level
called 'level' was also added to NachoPanel for testing.