Added NachoFrame and XP increment/decrement functions
Added a NachoFrame class which we can use for displaying out panels for
the game and the menu (when we make the menu). I also added
functionality in for increment and decrement xp. When an enemy makes it
past the screen your xp = xp -5. when you kill an enemy xp++. If player
xp is <= 0 the game ends. I also added the xp count in the stats corner
for testing purposes.